resultado final

+++ CartasSuperTrunfo.c
@@ -11,7 +11,7 @@ int main() {
     char estado1 = 'A', estado2 = 'B';
     char codigo1[] = "A01", codigo2[] = "B02";
     char cidade1[] = "Sao Paulo", cidade2[] = "Rio de Janeiro";
-    unsigned long int populacao1 = 12325000, populacao2 = 6748000;
+    int populacao1 = 12325000, populacao2 = 6748000;
     float area1 = 1521.11, area2 = 1200.25;
     float pib1 = 699.28, pib2 = 300.50;
     int pontosTuristicos1 = 50, pontosTuristicos2 = 30;
@@ -33,7 +33,7 @@ int main() {
     printf("Estado: %c\n", estado1);
     printf("Codigo: %s\n", codigo1);
     printf("Nome da Cidade: %s\n", cidade1);
-    printf("Populacao: %lu\n", populacao1); // População com %lu para unsigned long int
+    printf("Populacao: %d\n", populacao1); 
     printf("Area: %.2f km²\n", area1);
     printf("PIB: %.2f bilhoes de reais\n", pib1);
     printf("Numero de Pontos Turisticos: %d\n", pontosTuristicos1);
@@ -45,7 +45,7 @@ int main() {
     printf("Estado: %c\n", estado2);
     printf("Codigo: %s\n", codigo2);
     printf("Nome da Cidade: %s\n", cidade2);
-    printf("Populacao: %lu\n", populacao2); // População com %lu para unsigned long int
+    printf("Populacao: %d\n", populacao2); 
     printf("Area: %.2f km²\n", area2);
     printf("PIB: %.2f bilhoes de reais\n", pib2);
     printf("Numero de Pontos Turisticos: %d\n", pontosTuristicos2);
